created folder for utility components. moved randomizer and timer to utility folder and fixed imports.
